seithi-brain/requirements.txt
# Core
newspaper3k>=0.2.8
feedparser>=6.0.10
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9

# ML / AI
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
setfit>=1.0.3

# Utilities
tqdm>=4.66.1
requests>=2.31.0
numpy>=1.24.0
schedule>=1.2.0
lxml>=4.9.0,<5.2.0
